Move deploy tools to filesystems
